Update TypeScript configs and revamp API end-points
This commit overhauls the TypeScript configurations, refactors API end-points, and enhances the user controller with CRUD operations. It also includes the addition of new axios-based API files, changes in database host configuration, and the creation of router and middleware for request handling. Adjustment in the frontend components to reflect these backend changes is also conducted.